+++ stepup/core/nglob.py
@@ -75,7 +75,7 @@ It has the following use cases:
 
     - `feedback_001.md` with `report_123.pdf`
 
-- Conventional glob wildcards are also allowed and are called "anonymous wildcards"
+- Conventional (recursive) glob wildcards are also allowed and are called "anonymous wildcards"
   to clarify the distinction from named wildcards.
 """
 
@@ -103,6 +103,22 @@ __all__ = (
 
 @attrs.define
 class NGlobMatch:
+    """A set of matches corresponding sharing consistent values for named wildcards.
+
+    The matching files can accessed by integer indexing or through the `files` attribute:
+
+    ```python
+    assert match[0] == match.files[0]
+    ```
+
+    The substring matching the named wildcards can be accessed as attributes.
+    For example, the substring matching a named wildcard `foo` is accessed as follows:
+
+    ```python
+    print(match.foo)
+    ```
+    """
+
     _mapping: dict[str, str]
     _files: list[Path | list[Path]]
 
@@ -113,14 +129,23 @@ class NGlobMatch:
         try:
             return self._mapping[name]
         except KeyError as exc:
-            raise AttributeError(f"'NGlobMatch' object has not attribute '{name}'") from exc
+            raise AttributeError(f"'NGlobMatch' object has no attribute '{name}'") from exc
 
     @property
     def mapping(self) -> dict[str, str]:
+        """Dictionary with `(wildcard_name, substring)` items."""
         return self._mapping
 
     @property
-    def files(self) -> list[Path]:
+    def files(self) -> list[Path | list[Path]]:
+        """Matching files, all having consistent substrings matching the named wildcards.
+
+        Each item corresponds to the a pattern in `NGlobMulti.patterns`.
+        If a pattern has anonymous wildcards,
+        the item itself is a list of all files matching the pattern,
+        If the pattern contains no anonymous wildcards,
+        the corresponding item in the returned list is a single path.
+        """
         return self._files
 
 
@@ -137,12 +162,14 @@ class NGlobSingle:
 
     @classmethod
     def structure(cls, state, strings: list[str]):
+        """Create an instance from the result of the `unstructure` method."""
         results = {
             tuple(values): {Path(strings[path]) for path in paths} for values, paths in state["r"]
         }
         return cls(strings[state["p"]], state["s"], results)
 
     def unstructure(self, lookup: dict[str, int]) -> dict[str, Any]:
+        """Return a serializable representation of the instance."""
         results = [
             (key, [lookup[path] for path in sorted(paths)]) for key, paths in self._results.items()
         ]
@@ -162,31 +189,54 @@ class NGlobSingle:
 
     @property
     def pattern(self) -> str:
+        """The Named Glob pattern used to match filenames."""
         return self._pattern
 
     @property
     def subs(self) -> dict[str, str]:
+        """User-defined glob patterns for the named wildcards.
+
+        When a name is not present, `*` is used.
+        """
         return self._subs
 
     @property
     def results(self) -> dict[tuple[str, ...], set[Path]]:
+        """All matching files, grouped by substrings matching the named wildcards.
+
+        The keys of the `results` dictionary are tuples with the substrings,
+        matching the respective named wildcards in the `used_names` tuple.
+        The values are sets with matching paths.
+        """
         return self._results
 
     @property
     def used_names(self) -> tuple[str, ...]:
+        """A tuple of named wildcards present in the pattern."""
         return self._used_names
 
     @property
     def glob_pattern(self) -> str:
+        """The conversion of the named glob to a (more general) conventional glob pattern."""
         return self._glob_pattern
 
     @property
     def regex(self) -> re.Pattern:
+        """The conversion of the named glob to a regular expression."""
         return self._regex
 
     def _loop_matches(
         self, paths: Iterable[str]
     ) -> Iterator[tuple[tuple[str, ...], set[Path], Path]]:
+        """Low-level iterator used by the `extend` and `reduce` methods.
+
+        The paths are tested one by one against the regular expression.
+        In case of a hit, it yields a tuple with the following three items:
+
+        - `values`: the substrings matching the named wildcards.
+        - `path_set`: the current set of paths associated with the combination of substrings.
+        - `path`: a `Path` instance of the matching path.
+        """
         for path in paths:
             match_ = self._regex.fullmatch(path)
             if match_ is not None:
@@ -201,14 +251,28 @@ class NGlobSingle:
                     del self._results[values]
 
     def extend(self, paths: Iterable[str]) -> Iterator[tuple[str, ...]]:
-        """Add matching paths from the given list paths. Yield new values for named wildcards."""
+        """Add matching paths from the given list paths.
+
+        Yields
+        ------
+        values
+            A tuple with substring matching the named wildcards,
+            only this combination of names was not present yet.
+        """
         for values, path_set, path in self._loop_matches(paths):
             if len(path_set) == 0:
                 yield values
             path_set.add(path)
 
     def reduce(self, paths: Iterable[str]) -> Iterator[tuple[str, ...]]:
-        """Remove matching paths from given list paths. Yield deleted values of named wildcards."""
+        """Remove matching paths from given list paths.
+
+        Yields
+        ------
+        values
+            A tuple with deleted substring matching the named wildcards,
+            only if the last matching paths were removed.
+        """
         for values, path_set, path in self._loop_matches(paths):
             if len(path_set) > 0:
                 path_set.discard(path)
@@ -216,6 +280,14 @@ class NGlobSingle:
                     yield values
 
     def glob(self) -> Iterator[tuple[str, ...]]:
+        """Extend the results with paths obtained through the built-in glob module.
+
+        Yields
+        ------
+        values
+            A tuple with substring matching the named wildcards,
+            only this combination of names was not present yet.
+        """
         yield from self.extend(glob.glob(self._glob_pattern, recursive=True))
 
 
@@ -233,6 +305,7 @@ def has_anonymous_wildcards(pattern: str) -> bool:
 
 
 def iter_wildcard_names(pattern: str) -> Iterator[str]:
+    """Iterate over the names of the named wildcards in a Named Glob pattern."""
     for ipart, part in enumerate(RE_NAMED_WILD.split(pattern)):
         if ipart % 2 == 1 and part.startswith("${*"):
             yield part[3:-1]
@@ -240,6 +313,8 @@ def iter_wildcard_names(pattern: str) -> Iterator[str]:
 
 @attrs.define
 class NGlobMulti:
+    """A sequence of Named Glob patterns for which consistent matches are collected."""
+
     _nglob_singles: tuple[NGlobSingle, ...] = attrs.field()
     _subs: dict[str, str] = attrs.field(init=False)
     _used_names: tuple[str, ...] = attrs.field(init=False)
@@ -280,6 +355,7 @@ class NGlobMulti:
 
     @classmethod
     def structure(cls, state: list, strings: list[str]) -> Self:
+        """Create an instance from the result of the `unstructure` method."""
         subs = state[1].copy()
         for ngs_data in state[0]:
             ngs_data["s"] = subs
@@ -291,6 +367,7 @@ class NGlobMulti:
         return result
 
     def unstructure(self, lookup: dict[str, int]) -> list:
+        """Return a serializable representation of the instance."""
         data = [
             [ngs.unstructure(lookup) for ngs in self._nglob_singles],
             self._subs.copy(),
@@ -302,6 +379,18 @@ class NGlobMulti:
 
     @classmethod
     def from_patterns(cls, patterns: Iterable[str], subs: dict[str, str] | None = None) -> Self:
+        """Create a new instance for given patterns without any results.
+
+        Parameters
+        ----------
+        patterns
+            Named Glob patterns.
+            Results will be constrained to have consistently matching substrings
+            for the named wildcards appearing in all the patterns.
+        subs
+            Optional anonymous glob patterns for the named patterns.
+            When a name is not present, the wildcard `*` is used for this name.
+        """
         if isinstance(patterns, str):
             raise TypeError("The patterns argument cannot be a string")
         if not all(isinstance(pattern, str) for pattern in patterns):
@@ -317,63 +406,129 @@ class NGlobMulti:
 
     @property
     def nglob_singles(self) -> tuple[NGlobSingle, ...]:
+        """The list of NGlobSingle instances, one for each pattern.
+
+        These instances collect (partial) matches before any consistency is imposed between
+        the substrings matching the same name in different patterns.
+        """
         return self._nglob_singles
 
     @property
     def patterns(self):
+        """The list of Named Glob patterns."""
         return [ngs.pattern for ngs in self._nglob_singles]
 
     @property
     def subs(self) -> dict[str, str]:
+        """User-defined glob patterns for the named wildcards.
+
+        When a name is not present, `*` is used.
+        """
         return self._subs
 
     @property
     def used_names(self) -> tuple[str, ...]:
+        """The names used across all the named glob patterns."""
         return self._used_names
 
     @property
     def has_wildcards(self) -> bool:
+        """True if any named or anonymous wildcards are present in the patterns."""
         return self._has_wildcards
 
     @property
     def results(self) -> dict[tuple[str, ...], list[set[Path]]]:
+        """A dictionary with all matches collected so far.
+
+        A key in this dictionary is a tuple of substrings named wildcards,
+        using the same order as the `used_names` attribute.
+
+        A value is a list of sets of paths.
+        Each item in the list is a set of matching filenames for the corresponding
+        pattern from the `patterns` attribute, whose named wildcards match the substrings
+        of the key.
+
+        The results can be extended with the `extend` and `glob` methods.
+        Conversely, results can be removed with the `reduce` method.
+        """
         return self._results
 
-    def _iter_consistent(self, criteria: dict[str, str], full_paths: list | int):
+    def _iter_consistent(
+        self, criteria: dict[str, str], full_paths: list | int
+    ) -> Iterator[tuple[str, ...], list[list[Path]]]:
+        """Iterate over (partial) matching substrings and corresponding paths.
+
+        Parameters
+        ----------
+        criteria
+            A dictionary mapping named wildcards to matching substrings.
+        full_paths
+            If this is a list, it contains lists of paths matching the patterns
+            in of the `patterns` attribute with substrings consist to those in
+            the criteria argument.
+            Note that this is a recursive iterator, so full_paths may contain less
+            items than there are patterns when the recursion has not reached it full
+            depth yet.
+            If this is an integer, it is in index referring to the item in the `patterns`
+            to identify the current pattern being processed.
+        """
         start = full_paths if isinstance(full_paths, int) else len(full_paths)
         if start == len(self._nglob_singles):
+            # We're in the deepest recursion: yield a result.
             yield tuple(criteria[name] for name in self._used_names), full_paths
         else:
+            # Recursion in progress...
             ngs = self._nglob_singles[start]
             for new_values, paths in ngs.results.items():
                 next_criteria = criteria.copy()
+                # Check if named wildcards are consistent with the matching paths so far.
                 for name, new_value in zip(ngs.used_names, new_values, strict=False):
                     value = next_criteria.get(name)
                     if value is None:
                         next_criteria[name] = new_value
                     elif value != new_value:
+                        # Inconsistent matches for named wildcards in different patterns.
+                        # This cannot produce a useful result.
                         next_criteria = None
                         break
                 if next_criteria is not None:
+                    # Consistency can still be imposed, so enter the next recursion...
                     next_full_paths = (
                         start + 1 if isinstance(full_paths, int) else [*full_paths, paths]
                     )
                     yield from self._iter_consistent(next_criteria, next_full_paths)
 
     def _extend_consistent(self, i: int, values: tuple[str, ...]):
+        """Extend the results of this instance, given an added combination of matching substrings.
+
+        Parameters
+        ----------
+        i
+            The integer index of the pattern in the `patterns` attribute being processed.
+        values
+            A new set of substrings matching the named wildcards.
+        """
         criteria = dict(zip(self._nglob_singles[i].used_names, values, strict=False))
         new_items = list(self._iter_consistent(criteria, []))
         for full_values, full_paths in new_items:
             self._results[full_values] = full_paths
 
     def _reduce_consistent(self, i: int, values: tuple[str, ...]):
+        """Return the results of this instance, given a removed combination of matching substrings.
+        Parameters
+        ----------
+        i
+            The integer index of the pattern in the `patterns` attribute being processed.
+        values
+            A new set of substrings matching the named wildcards.
+        """
         criteria = dict(zip(self._nglob_singles[i].used_names, values, strict=False))
         old_items = list(self._iter_consistent(criteria, 0))
         for full_values, _ in old_items:
             del self._results[full_values]
 
     def extend(self, paths: Iterable[str]):
-        """Find matches from a list of paths. Return new full matches (and track partial matches)"""
+        """Try to extend the results by searching for matches in the given list of paths."""
         if isinstance(paths, str):
             raise TypeError("The paths argument cannot be a string.")
         for i, ngs in enumerate(self._nglob_singles):
@@ -381,7 +536,7 @@ class NGlobMulti:
                 self._extend_consistent(i, values)
 
     def reduce(self, paths: Iterable[str]):
-        """Drop matches based on deleted paths. Return old full matches (and track partial)."""
+        """Drop results by eliminating the provided paths."""
         if isinstance(paths, str):
             raise TypeError("The paths argument cannot be a string.")
         for i, ngs in enumerate(self._nglob_singles):
@@ -389,6 +544,7 @@ class NGlobMulti:
                 self._reduce_consistent(i, values)
 
     def glob(self):
+        """Extend the results with paths found by the built-in glob function."""
         for i, ngs in enumerate(self._nglob_singles):
             for values in ngs.glob():
                 self._extend_consistent(i, values)
@@ -404,7 +560,16 @@ class NGlobMulti:
     # Convenience methods
 
     def matches(self) -> Iterator[NGlobMatch]:
-        """Iterate over combinations of files that consistently match of all patterns."""
+        """Iterate over combinations of files that consistently match all patterns.
+
+        This offers a more convenient interface of the `results` attribute.
+
+        Yields
+        ------
+        nglob_match
+            An instance of NGlobMatch, which contains the substrings matching the named wildcards
+            and the corresponding lists of paths.
+        """
         for values, path_sets in sorted(self._results.items()):
             mapping = dict(zip(self._used_names, values, strict=False))
             files = [
@@ -426,10 +591,11 @@ class NGlobMulti:
         return tuple(sorted(result))
 
     def __bool__(self):
-        """True when there were some matches."""
+        """True when there are some items in the `results` attribute."""
         return len(self.results) > 0
 
     def __iter__(self) -> Iterator[str | NGlobMatch]:
+        """Iterates over `self.matches` if there are named wildcards, else over `self.files`."""
         if len(self._used_names) > 0:
             return self.matches()
         else:
@@ -444,37 +610,37 @@ class NGlobMulti:
         """
         return any(ngs.regex.fullmatch(path) for ngs in self._nglob_singles)
 
-    def may_change(self, deleted: set[str], added: set[str]) -> bool:
+    def may_change(self, deleted: set[str], updated: set[str]) -> bool:
         """Determine whether the results may change (later) after deleting or adding files.
 
         Parameters
         ----------
         deleted
             Set of files to be deleted.
-        added
-            Set of files to be added.
+        updated
+            Set of files to be updated.
 
         Returns
         -------
         may_change
             True if the NGlobMulti results may change.
             (It may require additional additions and deletions to get any effect,
-            but cannot be excluded that that is possible.)
+            but cannot be excluded that the provided deletions and updates play a role in it.)
         """
-        added_new = added.copy()
+        updated_new = updated.copy()
         for ngs in self._nglob_singles:
             for paths in ngs.results.values():
                 if not deleted.isdisjoint(paths):
                     return True
-                added_new.difference_update(paths)
+                updated_new.difference_update(paths)
         for ngs in self._nglob_singles:
-            for path in added_new:
+            for path in updated_new:
                 if ngs.regex.fullmatch(path):
                     return True
         return False
 
     def will_change(self, deleted: Collection[str], updated: Collection[str]) -> bool:
-        """Determine whether the results may change after deleting or adding files.
+        """Determine whether the results will change after deleting or adding files.
 
         Parameters
         ----------
@@ -505,16 +671,17 @@ def convert_nglob_to_regex(
         A string with named wildcards.
     subs
         A dictionary mapping names to glob patterns.
-        If a name is not present, "*" is used as default.
+        If a name is not present, `*` is used as default.
     allow_names
-        When set to False, named wildcards are not allowed.
+        When set to `False`, named wildcards are not allowed.
 
     Returns
     -------
     regex
-        A regex string to test if a string matches the pattern.
+        A regular expression string to test if a string matches the pattern.
         It also contains symbolic groups to extract values
-        corresponding to named wildcards.
+        corresponding to named wildcards
+        and to impose consistency when the same name appears multiple times.
     """
     if subs is None:
         subs = {}
@@ -551,7 +718,7 @@ def convert_nglob_to_regex(
 
 
 def convert_nglob_to_glob(pattern: str, subs: dict[str, str] | None = None) -> str:
-    """Convert nglob wildcards to ordinary ones, compatible with builtin glob and fnmatch.
+    """Convert nglob wildcards to ordinary ones, compatible with builtin glob and fnmatch modules.
 
     Parameters
     ----------
@@ -559,12 +726,13 @@ def convert_nglob_to_glob(pattern: str, subs: dict[str, str] | None = None) -> s
         A string with named wildcards.
     subs
         A dictionary mapping names to glob patterns.
-        If a name is not present, "*" is used as default.
+        If a name is not present, `*` is used as default.
 
     Returns
     -------
+    pattern
         A conventional wildcard string, without the constraint that named wildcards must correspond.
-        Where possible, neighbouring wildcards are merged into one.
+        Where possible, neighboring wildcards are merged into one.
     """
     if subs is None:
         subs = {}